# src/main/java/com/telcaria/rc/nbi/wrappers/InfrastructureDay2ConfigurationWrapper.java
package com.telcaria.rc.nbi.wrappers;

import java.util.List;
import lombok.Data;

@Data
public class InfrastructureDay2ConfigurationWrapper {

  List<InfrastructureMetricWrapper> infrastructureMetricsInfo;

}